refactor: RequestHandler로 책임 분리
