+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -6,6 +6,7 @@ import java.net.Socket;
 import java.nio.file.Files;
 import nextstep.jwp.ContentType;
 import nextstep.jwp.HttpRequest;
+import nextstep.jwp.RequestHandler;
 import nextstep.jwp.exception.UncheckedServletException;
 import org.apache.coyote.Processor;
 import org.slf4j.Logger;
@@ -32,35 +33,11 @@ public class Http11Processor implements Runnable, Processor {
         try (final var inputStream = connection.getInputStream();
             final var outputStream = connection.getOutputStream()) {
             HttpRequest request = new HttpRequest(inputStream);
-            if (request.getTarget().equals("/")) {
-                final var response = getResponse("Hello world!", ContentType.HTML);
-                outputStream.write(response.getBytes());
-                outputStream.flush();
-            } else {
-                String fileUrl = "static" + request.getTarget();
-                File file = new File(
-                    getClass()
-                        .getClassLoader()
-                        .getResource(fileUrl)
-                        .getFile()
-                );
-                String responseBody = new String(Files.readAllBytes(file.toPath()));
-                String response = getResponse(responseBody, ContentType.from(file.getName()));
-                outputStream.write(response.getBytes());
-                outputStream.flush();
-            }
-
+            String response = RequestHandler.handle(request);
+            outputStream.write(response.getBytes());
+            outputStream.flush();
         } catch (IOException | UncheckedServletException e) {
             log.error(e.getMessage(), e);
         }
-    }
-
-    private static String getResponse(String responseBody, ContentType contentType) {
-        return String.join("\r\n",
-            "HTTP/1.1 200 OK ",
-            "Content-Type: "+ contentType.value  + ";charset=utf-8 ",
-            "Content-Length: " + responseBody.getBytes().length + " ",
-                "",
-            responseBody);
     }
 }
